requirements.txt
-f https://www.paddlepaddle.org.cn/whl/linux/mkl/avx/stable.html
paddlepaddle-gpu==2.3.2.post101
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cu101
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.8.2+cu101
-f https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.7.2
paddlenlp==2.5.2
transformers==4.30.1
sentencepiece==0.1.99
fugashi==1.3.0
ipadic==1.0.0
lightgbm
scikit-learn==1.3.2
yacs
seqeval
pypandoc
attrdict3
python_docx
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy
visualdl
rapidfuzz
opencv-python<=4.6.0.66
opencv-contrib-python<=4.6.0.66
cython
lxml
premailer
openpyxl
attrdict
PyMuPDF<1.21.0
Pillow<=9.5.0